Fix `repo download project nn/p` formatting
The download command wants change-id/patchset not refs/changes/ ref
style string as used by git fetch or pull. Break off the last two
path segments and pass this to repo download.

Move the repo download part of the command name into RepoCommand,
leaving RepoScheme to just return the project name. This better
matches the other types of commands and schemes where the scheme
names the project and the command supplies the invocation and
inserts the URL at the proper position.

Change-Id: I0473b61cccbc3bc1ef38c39aeaa4057ada63f3cc