Indicate whether each download scheme supports authentication
isAuthSupported() defaults to isAuthRequired(), which is the value
used here, so no behavior change intended.

Explicitly specifying whether authentication is supported makes it
clearer which download schemes should be shown for changes that are
not visible to anonymous users.  Later we can stop providing a default
implementation in DownloadScheme to prevent new download schemes from
getting a wrong setting by forgetting to override the default.

Change-Id: If6f11bf906c58c721b0a9c39b9dbee48384b6e32