Always show repo download command if repo download scheme is enabled
At the moment the repo download command is only shown in the repo
download scheme if it is explicitly enabled in the gerrit.config by
setting download.command = repo_download. Since the repo download
command is the only command that makes sense for the repo download
scheme it should always be displayed when the repo download scheme is
enabled. It doesn't make sense to be able to enable the repo download
scheme and to disable the repo download command at the same time.

Change-Id: I79758f81c43b43c52dc6ecd116b7cda66f02374b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>