Return null as URL for SshScheme if user has no username
For SSH a username is required and returning null in this case is
better than return an invalid URL. A null URL is already handled as
'scheme not available'.

At the moment in case of SSH the client makes an extra check whether
the user has an username and if not the SSH scheme is skipped.

Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>
Change-Id: I827af45929ea34a490f52658c67a13cd4591dd27