Implement clone commands
At the moment the clone commands are built on Gerrit client-side. Now
this plugin provides these clone commands via the new CloneCommand
extension point and in a next step the Gerrit client can retrieve the
clone commands via the /config/server/info REST endpoint so that the
client code to build the clone commands can be removed.

Change-Id: I2f284b3995ebe3ab17df331713be3d76c9e08dd6
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>