[RFC] Use commit IDs for download commands when change refs are hidden
Git has a configuration option to hide refs from the initial
advertisement (uploadpack.hideRefs). This option can be used to hide
the change refs from the client. As consequence fetching changes by
change ref is not working anymore. However by setting
uploadpack.allowTipSha1InWant to true fetching changes by commit ID is
possible. Adapt the git download commands so that they use the commit
ID instead of the change ref when a project is configured like this.
If tools use the download commands that are returned by the Gerrit
server, switching on this configuration should be transparent for
them.

Bug: Issue 175
Change-Id: I0f1f68c856f23ffeddd0e833bf45b31549cfd6f2
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>