Update patch set 1
Patch Set 1:

> Would this break integrations to CI like Jenkins though?
 > They rely on the stream events on the advertised change to fetch
 > them directly from Gerrit using the refs/changes/* syntax.

I guess it will break the Jenkins integration. But then the Jenkins side should be fixed. Clients should not construct the download commands on their own, but retrieve the download commands from the Gerrit server. We may add a possibility to have the download commands included into the stream-events if this makes things easier. In any case this change only changes the download commands *if* you have configured your project to hide the change refs. If you have done this, the Jenkins integration is anyway broken, independently of this change. So this change doesn't make things worse. If you rely on the Jenkins integration you should not set this configuration on your project :-) Still projects which do not use Jenkins may benefit from it.

Patch-set: 1