Refactor repo commands
We've had a single "repo command" stuffed in the download UI which
doesn't make a lot of sense.  The command types are more about how
they mutate your local checkout, not the commands used to download.

Lets merge the repo-vs-git logic into GitDownloadCommand directly.
We create separate "getRepoCommand" and "getCommand" methods for
downstream classes to extend with the default "getRepoCommand"
defaulting to null.

Bug: Issue 11594
Change-Id: I6848c431975ec4db3aa09a160af5198b93ce5bb9