refactor repo commands
We've had a single "repo command" stuffed in the download UI which
doesn't make a lot of sense.  The command types are more about how
they mutate your local checkout, not the commands used to download.

Lets create a new RepoDownloadCommand interface akin to the existing
GitDownloadCommand, and then merge the RepoCommand into the existing
CheckoutCommand since that's what the `repo download` is doing.

Bug: https://crbug.com/gerrit/11594
Change-Id: Ie6d35f27290d504e72bc8e1de3d6adb581027985