New command: "Rebase On"
Add a new command that allows to rebase the currently checked out
branch onto the fetched commit. This is useful in the case where
a user has one or more commits on a local branch and needs to apply
them on top of commits that another user has pushed for review. It
avoids having to use one of the other existing commands (usually
the checkout command) and then manually go back to the previously
checked out branch and invoke the rebase.

Change-Id: I505cbd57d5645bf2f9dcdc5e2309446462a15563