repo: include patchset number in commands
Since we don't include the ps number, repo will look up the latest
one and use that.  If the user is viewing an older ps, that means
the repo command won't actually download the version they have
selected which can be quite confusing.

The git version already includes the fully change id + ps num, so
this brings repo up to parity.

Change-Id: I6a4cd56cf9960d626770f98837028119aeaed4d4