Only attempt to make hook executable if it downloaded
There is no reason to attempt to chmod the commit-msg hook if it failed
to download, so use '&&' instead of ';'

Some shells, and some shell extensions in particular (f.i. zsh and
oh-my-zsh), have url magic features that automatically attempt to
escape characters in conjunction with urls.

This caused the current implementation to fail for some of our users as
it would escape the ';' because it was connected to the curl command
argument.

This resolves both issues by chaining the command instead.

Change-Id: I4304abbb11b7be63686bec51b200874fa00eab81