Use fast-forward when using git cherry-pick.
Ideally, when extending work on an existing non-merged patchset
from a colleague on gerrit, one starts with the "branch" download
command. Then, when uploading a new patchset to gerrit, this
will not lead to a rebase of the existing patchsets.

However, sometimes users use cherry-pick for extending work.
Use fast-forward, such that the git-commit is preserved if
possible. This reduces the numbers of cases when a rebase
of the existing patchset is required.

Change-Id: I72c8a6464fa074a478ed6cc345a99273da671589