SshScheme: Provide an option to omit usernames in SSH URLs
Including the username in the download URL can prevent IT admins from
configuring hosts to redirect read-only traffic to regional replicas
when the admins rely on the git url.insteadOf configuration. This is
because git doesn't support anything except simple replacement of
matching strings, so there's no way to account for variable usernames in
a host-wide /etc/gitconfig file.

The default stays as always including the username in the SSH download
commands.

Release-Notes: Added an option to omit SSH usernames from download commands
Change-Id: I49f3b85adc3dae2d440366c919fb656ca95b812f