Clone: configure pushUrl if sshdAdvertisedPrimaryAddress is set
The option plugin.download-commands.advertisedPrimaryAddress specifies
the address clients should be configured to reach a primary Gerrit
instance.

This may differ from sshd.listenAddress if fetch is served from
another address. An example is a setup where upload-pack requests
are served by a Gerrit replica and receive-pack by a Gerrit primary.
Since ssh cannot be load balanced on layer 7 the addresses of the
primary and replica need to be different.

Configure remote.origin.pushUrl in the clone command installing the
commit-msg hook which is typically used if the client is used to push
new changes.

Change-Id: I69a271c55fe53a4d1c01b5f24f1a6fd8313e7ba8