Add curl '-f' option to fail on download problems
By default curl will exit with exit code '0' even if there
was a problem with the transfer. Option '-f' will make curl
to exit with error code '22' if HTTP response code was 400
or higher. If there was a problem with transfer, it does not
make sense to process the next command - chmod.
Option '-f' was introduced in curl in version 4.0, released
in 1998, so is safe to be used.

Change-Id: I74b1669b48d10a690a290dae2c24480f3a6994c6